+++ roomserver/src/room/character.rs
@@ -32,7 +32,7 @@ impl Into<CharacterPt> for Character {
 ///段位数据
 #[derive(Clone, Debug)]
 pub struct League {
-    pub league_id: u8,    //段位id
+    pub league_id: i8,    //段位id
     pub score: i32,       //段位积分
     pub league_time: i64, //进入段位时间
 }
@@ -40,9 +40,9 @@ pub struct League {
 impl From<&LeaguePt> for League {
     fn from(lp: &LeaguePt) -> Self {
         let mut l = League::default();
-        l.score = lp.get_league_score() as i32;
+        l.score = lp.get_league_score();
         l.league_time = lp.league_time;
-        let league_id = lp.league_id as u8;
+        let league_id = lp.league_id as i8;
         l.league_id = league_id;
         l
     }
@@ -69,13 +69,13 @@ impl League {
 
     pub fn into_pt(&self) -> LeaguePt {
         let mut lpt = LeaguePt::new();
-        lpt.set_league_id(self.league_id as u32);
+        lpt.set_league_id(self.league_id as i32);
         lpt.set_league_score(self.score);
         lpt.set_league_time(self.league_time);
         lpt
     }
 
-    pub fn update(&mut self, league_id: u8, league_score: i32, league_time: i64) {
+    pub fn update(&mut self, league_id: i8, league_score: i32, league_time: i64) {
         self.league_id = league_id;
         self.score = league_score;
         self.league_time = league_time;
